+++ src/wai/annotations/roi_utils/_ROIObject.py
@@ -1,14 +1,38 @@
-from typing import Dict, Tuple
+from typing import Dict, List, Optional, Set
+
+from ..core.logging import LoggingEnabled
 
 
-class ROIObject:
+class ROIObject(LoggingEnabled):
     """
     Holds a detected object from a ROI file.
     """
+    logger_name = "wai.annotations.ROIObject"
+
+    # The ordered required arguments to ROIObject
+    required_keywords = ("x0", "y0", "x1", "y1",
+                         "x0n", "y0n", "x1n", "y1n",
+                         "label", "label_str", "score")
+
+    # The ordered optional arguments to ROIObject
+    optional_keywords = ("poly_x", "poly_y", "poly_xn", "poly_yn",
+                         "minrect_w", "minrect_h")
+
+    # The ordered keyword arguments to ROIObject
+    keywords = required_keywords + optional_keywords
+
+    # Cached set representations of keywords for fast lookups
+    required_keyword_set: Set[str] = set(required_keywords)
+    optional_keyword_set: Set[str] = set(optional_keywords)
+    keyword_set: Set[str] = set(keywords)
+
     def __init__(self,
                  x0: float, y0: float, x1: float, y1: float,
                  x0n: float, y0n: float, x1n: float, y1n: float,
-                 label: int, label_str: str, score: float):
+                 label: int, label_str: str, score: float,
+                 poly_x: Optional[List[float]] = None, poly_y: Optional[List[float]] = None,
+                 poly_xn: Optional[List[float]] = None, poly_yn: Optional[List[float]] = None,
+                 minrect_w: Optional[float] = None, minrect_h: Optional[float] = None):
         self.x0: float = x0
         self.y0: float = y0
         self.x1: float = x1
@@ -20,6 +44,37 @@ class ROIObject:
         self.label: int = label
         self.label_str: str = label_str
         self.score: float = score
+        self.poly_x: Optional[List[float]] = poly_x
+        self.poly_y: Optional[List[float]] = poly_y
+        self.poly_xn: Optional[List[float]] = poly_xn
+        self.poly_yn: Optional[List[float]] = poly_yn
+        self.minrect_w: Optional[float] = minrect_w
+        self.minrect_h: Optional[float] = minrect_h
+
+        # Make sure all of the polygon lists are the same length
+        self.ensure_polygon_list_lengths()
+
+    def ensure_polygon_list_lengths(self):
+        """
+        Makes sure all polygon lists match in length.
+        """
+        # Get the length of each list
+        len_x = -1 if self.poly_x is None else len(self.poly_x)
+        len_y = -1 if self.poly_y is None else len(self.poly_y)
+        len_xn = -1 if self.poly_xn is None else len(self.poly_xn)
+        len_yn = -1 if self.poly_yn is None else len(self.poly_yn)
+
+        # Make sure all lengths are the same
+        if len_x != len_y or len_y != len_xn or len_xn != len_yn:
+            raise ValueError("Polygon lists have differing lengths")
+
+    def has_polygon(self) -> bool:
+        """
+        Whether this ROI object has a polygon.
+
+        :return:    True if the object has a polygon.
+        """
+        return self.poly_x is not None
 
     def image_width(self) -> int:
         """
@@ -43,32 +98,32 @@ class ROIObject:
         else:
             return round(self.y1 / self.y1n)
 
-    @classmethod
-    def keywords(cls) -> Tuple[str, ...]:
-        """
-        Gets the names of the keyword parameters to this class.
-
-        :return:    The keyword parameter names.
-        """
-        return "x0", "y0", "x1", "y1", "x0n", "y0n", "x1n", "y1n", "label", "label_str", "score"
-
-    @classmethod
-    def types(cls) -> Tuple[type, ...]:
-        """
-        Gets the types of the keyword parameters to this class.
-
-        :return:    The keyword parameter types.
-        """
-        return float, float, float, float, float, float, float, float, int, str, float
-
     def as_dict(self) -> Dict[str, str]:
         """
         Gets a dictionary representation of this object.
 
         :return:    A dict.
         """
-        return {keyword: str(getattr(self, keyword))
-                for keyword in self.keywords()}
+        # Create a result dict
+        result = {}
+
+        # Process each keyword in turn
+        for keyword in self.keywords:
+            # Get the value for the attribute
+            value = getattr(self, keyword)
+
+            # Ignore missing optional attributes
+            if value is None:
+                continue
+
+            # Format the polygon lists as strings
+            if isinstance(value, list):
+                value = ",".join(map(str, value))
+
+            # Add the string representation of the value to the dict
+            result[keyword] = value
+
+        return result
 
     @classmethod
     def from_dict(cls, dict: Dict[str, str]) -> "ROIObject":
@@ -78,5 +133,43 @@ class ROIObject:
         :param dict:    The dict.
         :return:        A ROIObject instance.
         """
-        return ROIObject(**{keyword: type(dict[keyword])
-                            for keyword, type in zip(cls.keywords(), cls.types())})
+        # Create a set of keyword arguments
+        kwargs = {}
+
+        # Process each dictionary item in turn
+        for keyword, value in dict.items():
+            # Ignore unexpected keywords
+            if keyword not in cls.keyword_set:
+                # Log anything other than "file" which we know will be in there
+                if keyword != "file":
+                    cls.logger().info(f"Unexpected keyword '{keyword}' in ROI dictionary, ignoring")
+                continue
+
+            # Add the value to the kwargs
+            kwargs[keyword] = cls._parse(keyword, value)
+
+        return ROIObject(**kwargs)
+
+    @classmethod
+    def _parse(cls, keyword: str, value: str):
+        """
+        Parses the string representation of a value to its internal type.
+
+        :param keyword: The name of the attribute.
+        :param value:   The string representation of the attribute's value.
+        """
+        # Float types
+        if keyword in {"x0", "y0", "x1", "y1", "x0n", "y0n", "x1n", "y1n", "score",
+                       "minrect_w", "minrect_h"}:
+            return float(value) if value is not "" else None
+
+        # Int types
+        if keyword in {"label"}:
+            return int(value)
+
+        # List of float types
+        if keyword in {"poly_x", "poly_y", "poly_xn", "poly_yn"}:
+            return list(map(float, value.split(","))) if value is not "" else None
+
+        # All other attributes are strings
+        return value