+++ src/wai/annotations/roi_utils/__init__.py
@@ -1,4 +1,8 @@
 """
 Package of utilities for working with the ROI format.
 """
+from ._combine_dicts import combine_dicts
+from ._min_rect_from_roi_polygon import min_rect_from_roi_polygon
+from ._polygon_from_roi_object import polygon_from_roi_object
+from ._roi_polygon import roi_polygon
 from ._ROIObject import ROIObject